refactor(Distance): 값 타입을 double로 수정

+++ src/main/java/subway/domain/Distance.java
@@ -6,14 +6,14 @@ public class Distance {
 
     private static final int MIN_VALUE = 0;
 
-    private final int value;
+    private final double value;
 
-    public Distance(final int value) {
+    public Distance(final double value) {
         validatePositive(value);
         this.value = value;
     }
 
-    private void validatePositive(final int value) {
+    private void validatePositive(final double value) {
         if (value < MIN_VALUE) {
             throw new IllegalArgumentException("거리는 음수가 될 수 없습니다.");
         }
